Moved services which are related to the report to a dedicated directory
